Default values for multiple devices setup in app.
