+++ GridConfiguration/src/main/java/com/sahu/gridconfiguration/Configuration.kt
@@ -4,11 +4,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.compositionLocalOf
 import androidx.compose.runtime.neverEqualPolicy
 import androidx.compose.runtime.remember
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.dimensionResource
-import androidx.compose.ui.res.integerResource
 import androidx.compose.ui.unit.Dp
-import androidx.compose.ui.unit.dp
 
 /**
  * Grid Configuration is the system that the screen is divided into multiple columns.
@@ -35,23 +31,6 @@ data class GridConfiguration(
     val gutterWidth: Dp,
     val totalColumns: Int,
 )
-
-/**
- * Empty params call takes the default gird system that is predefined in theme for
- * complete Device screen.
- *
- * @author Sahruday (SAHU)
- *
- * **int a, b = 83, 74**
- */
-@Composable
-fun rememberGridConfiguration(): GridConfiguration =
-    rememberGridConfiguration(
-        layoutWidth = LocalConfiguration.current.screenWidthDp.dp,
-        horizontalMargin = dimensionResource(id = R.dimen.layout_horizontal_margin),
-        gutterWidth = dimensionResource(id = R.dimen.layout_gutter_width),
-        totalColumns = integerResource(id = R.integer.layout_columns)
-    )
 
 /**
  * Customisable grid system to configure if required for the sub-layout which also
