+++ app/build.gradle
@@ -62,4 +62,6 @@ dependencies {
   androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
   androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
   debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
+
+  implementation(project(":GridConfiguration"))
 }
