+++ app/src/main/res/values/grid-system.xml
@@ -2,5 +2,5 @@
 <resources>
     <integer name="layout_columns">4</integer>
     <dimen name="layout_horizontal_margin">16dp</dimen>
-    <dimen name="layout_gutter_width">12dp</dimen>
+    <dimen name="layout_gutter_width">16dp</dimen>
 </resources>
